book_pipeline/requirements.txt
# Core dependencies
pymupdf>=1.22.0
numpy>=1.22.0
opencv-python>=4.7.0
pandas>=1.4.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-google-genai>=0.0.3
pydantic>=2.0.0
joblib>=1.2.0
scikit-learn >=1.0.0 
# Optional OCR dependencies
mistralai>=0.0.7  # For Mistral OCR

# API and web server
fastapi>=0.109.0
uvicorn>=0.27.0

# Monitoring and metrics
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing and development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
matplotlib>=3.5.0 
# Documentation generation